rename files with AttributeDistribution
